requirements.txt
Django
Pillow>=2.9.0
django-codemirror2
